Add files via upload
fixed a bug where the output folder needed to exist for FSL Bet to run correctly. On our test server this didn't trigger an error because things already existed in cache. Need to check if this is a problem for other notebooks as well @aswinnarayanan @iishiishii @MoniDoerig
Committed to old repo and new neurodel because I didn't know if we already switched over.